Added custom mirrored ore generation + Refactoring
