+++ map_gen/maps/space_race.lua
@@ -1,1 +1,1 @@
-return require 'map_gen.maps.space_race.map'
+return require 'map_gen.maps.space_race.map_gen.map'
