+++ map_gen/shared/generate.lua
@@ -317,6 +317,9 @@ end
 -- Warning! Changing this table after on_init or on_load has run will cause desyncs!
 -- @return dictionary of surface_name -> shape function
 function Public.get_surfaces()
+    if _LIFECYCLE == 8 then
+        error('Calling Generate.get_surfaces after on_init() or on_load() has run is a desync risk.', 2)
+    end
     return surfaces
 end
