+++ utils/token.lua
@@ -7,7 +7,7 @@ local counter = 0
 --- Assigns a unquie id for the given var.
 -- This function cannot be called after on_init() or on_load() has run as that is a desync risk.
 -- Typically this is used to register functions, so the id can be stored in the global table
--- instead of the function. This is becasue closures cannot be safely stored in the global table.
+-- instead of the function. This is because closures cannot be safely stored in the global table.
 -- @param  var<any>
 -- @return number the unique token for the variable.
 function Token.register(var)
